example-voting-app-sqlserver/cnab/app/voting-app/requirements.txt
Flask
Redis
azure.storage.blob